Experience in 4 bytes

+++ server/game/equations.h
@@ -55,15 +55,15 @@ public:
 //    const int eqGoldDrop(NPC& npc);
 
     // Limite de experiencia del nivel actual
-    const int eqExperienceLimit(Player& player);
+    const long eqExperienceLimit(Player& player);
 
     // Experiencia obtenida por ataque realizado
-    const int eqExperienceAttack(Player& player, Player& other);
-//    const int eqExperienceAttack(Player& player, NPC& npc);
+    const long eqExperienceAttack(Player& player, Player& other);
+//    const long eqExperienceAttack(Player& player, NPC& npc);
 
     // Experiencia obtenida por matar a un enemigo
-    const int eqExperienceKill(Player& player, Player& other);
-//    const int eqExperienceKill(Player& player, NPC& npc);
+    const long eqExperienceKill(Player& player, Player& other);
+//    const long eqExperienceKill(Player& player, NPC& npc);
 
     // Daño producido por ataque realizado
     const int eqAttackDamage(Player& player);
